cluster-bootstrap: make IsValidBootstrapToken() be in constant-time
The function uses BootstrapTokenRegexp.MatchString(token)
which is not a recommended practice.

Instead, break down the token into its components: ID, secret.
The ID is public thus we can use Regexp matching for it.
The secret needs constant time comparison. Iterate over
every character and make sure it fits the 0-9a-z range and that
it has a length of 16.

Kubernetes-commit: 1d519f1b0876f1c48cb9e08390064ea11f795be3